Split and move JS classes and functions into their own modules
Using a Singleton pattern for the `API`, `Controller`, `DOM` and `UI` classes, because these all require no more than a single instance.

Note that jQuery is also called in `.noConflict()` mode to be able to
use `$` inside functions without worrying about conflicts with other
libraries.
