+++ src/js/helpers/utils.js
@@ -1,0 +1,9 @@
+/**
+ * Util functions
+ */
+
+function getSlugFromURL( url ) {
+	return url.split( '/' ).pop();
+}
+
+export { getSlugFromURL };
